# Cleanups around message handling