--- a/benchmarks/token-parser/colmetadata-token.js
+++ b/benchmarks/token-parser/colmetadata-token.js
@@ -15,7 +15,7 @@ function main({ n, tokenCount }) {
   bench.start();
 
   for (let i = 0; i < n; i++) {
-    parser.addBuffer(data);
+    parser.write(data);
   }
 
   bench.end(n);