--- a/src/request.ts
+++ b/src/request.ts
@@ -273,6 +273,10 @@ class Request extends EventEmitter {
 
   on(event: 'cancel', listener: () => void): this
 
+  on(event: 'pause', listener: () => void): this
+
+  on(event: 'resume', listener: () => void): this
+
   on(event: string | symbol, listener: (...args: any[]) => void) {
     return super.on(event, listener);
   }
@@ -317,6 +321,14 @@ class Request extends EventEmitter {
    * @private
    */
   emit(event: 'cancel'): boolean
+  /**
+   * @private
+   */
+  emit(event: 'pause'): boolean
+  /**
+   * @private
+   */
+  emit(event: 'resume'): boolean
   /**
    * @private
    */
@@ -534,10 +546,8 @@ class Request extends EventEmitter {
     if (this.paused) {
       return;
     }
+    this.emit('pause');
     this.paused = true;
-    if (this.connection) {
-      this.connection.pauseRequest(this);
-    }
   }
 
   /**
@@ -549,9 +559,7 @@ class Request extends EventEmitter {
       return;
     }
     this.paused = false;
-    if (this.connection) {
-      this.connection.resumeRequest(this);
-    }
+    this.emit('resume');
   }
 
   /**