--- a/src/token/token.ts
+++ b/src/token/token.ts
@@ -492,12 +492,3 @@ export class SSPIToken extends Token {
     this.ntlmpacketBuffer = ntlmpacketBuffer;
   }
 }
-
-export class EndOfMessageToken extends Token {
-  declare name: 'EOM';
-  declare event: 'endOfMessage';
-
-  constructor() {
-    super('EOM', 'endOfMessage');
-  }
-}