This further cleans up some logic around message handling.

The biggest change here is that each message is handled by a new token parser object. This allows us to get rid of the "fake" `EndOfMessageToken`, and we can simply wait for the `end` event coming from the token parser. This opens up the possibility to use directly `.pipe` a message into the token parser directly, or convert the token parser to an async generator function that consumes a message. This will be explored in a follow up PR.

There's some additional changes around pause/resume and cancellation behavior, but those should not make much of a difference.

---

These changes have a small negative impact on performance, because each incoming message generates a new token parser object, and registers a bunch of event handlers. The performance hit is rather small, from what I've benchmarked:

#### Before

```
root ➜ /workspace (master ✗) $ node benchmarks/query/select-many-rows.js 
query/select-many-rows.js size=10 n=10: 588.2487546332677 (minor: 0 - 0ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=100 n=10: 261.9831279197856 (minor: 0 - 0ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=1000 n=10: 99.80497807868491 (minor: 3 - 2.60689ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=10000 n=10: 19.306442130170502 (minor: 27 - 15.837257999999999ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=10 n=100: 772.554154578383 (minor: 2 - 2.800477ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=100 n=100: 538.9699181054613 (minor: 4 - 3.93395ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=1000 n=100: 184.51681130638127 (minor: 29 - 14.130182999999999ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=10000 n=100: 23.826393117095794 (minor: 267 - 110.2691079999999ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=10 n=1000: 1,167.9108255236938 (minor: 11 - 13.298981999999999ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=100 n=1000: 793.9495864957719 (minor: 39 - 33.63353200000001ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=1000 n=1000: 202.59161794052602 (minor: 289 - 140.69072700000007ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=10000 n=1000: 24.29268440602386 (minor: 2666 - 1016.4431079999999ms, major: 0 - 0ms, incremental: 0 - 0ms)
```

#### After

```
root ➜ /workspace (arthur/cleanups ✗) $ node benchmarks/query/select-many-rows.js 
query/select-many-rows.js size=10 n=10: 593.4516292116595 (minor: 0 - 0ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=100 n=10: 283.3080613099043 (minor: 0 - 0ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=1000 n=10: 93.0445669146658 (minor: 3 - 3.511271ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=10000 n=10: 20.052101777089447 (minor: 27 - 12.085761ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=10 n=100: 791.370497247285 (minor: 2 - 2.444847ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=100 n=100: 522.5869254760995 (minor: 4 - 3.794459ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=1000 n=100: 178.18937654751682 (minor: 29 - 17.400887ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=10000 n=100: 23.107443947021952 (minor: 267 - 135.67801599999999ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=10 n=1000: 1,138.7238822945124 (minor: 12 - 16.9776ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=100 n=1000: 766.2731053515291 (minor: 40 - 44.07308800000001ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=1000 n=1000: 204.67734657150825 (minor: 290 - 159.481954ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=10000 n=1000: 23.797390898453546 (minor: 2667 - 1067.162521999999ms, major: 0 - 0ms, incremental: 0 - 0ms)
```